add gateway

+++ gateway/src/main/java/ru/practicum/shareit/booking/dto/BookItemRequestDto.java
@@ -2,6 +2,7 @@ package ru.practicum.shareit.booking.dto;
 
 import java.time.LocalDateTime;
 
+
 import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.FutureOrPresent;
 import lombok.AllArgsConstructor;
@@ -12,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BookItemRequestDto {
-	private long itemId;
+	private Integer itemId;
 	@FutureOrPresent
 	private LocalDateTime start;
 	@Future
